Move resource service fixtures to conftest
