+++ ofrak_core/test_ofrak/service/serialization_service/test_resource_service_serialization.py
@@ -1,53 +1,33 @@
 import pytest
 
 from ofrak.service.resource_service_i import ResourceServiceInterface
-from ofrak.service.serialization.serializers.serializer_i import SerializerInterface
+from ofrak.service.serialization.service_i import SerializationServiceInterface
 from test_ofrak.service.resource_service.test_resource_service import TestResourceService
-from test_ofrak.service.resource_service.test_resource_service import (
-    resource_service,  # noqa
-    tree1_resource_models,  # noqa
-    tree2_resource_models,  # noqa
-    tree3_resource_models,  # noqa
-    basic_populated_resource_service as _basic_populated_resource_service,
-    populated_resource_service as _populated_resource_service,
-    triple_populated_resource_service as _triple_populated_resource_service,
-)
 
 
 @pytest.fixture
 async def basic_populated_resource_service(
-    resource_service: ResourceServiceInterface,
-    tree1_resource_models,
-    serializer: SerializerInterface,
+    basic_populated_resource_service,
+    serializer: SerializationServiceInterface,
 ):
-    return _reserialize(
-        await _basic_populated_resource_service(resource_service, tree1_resource_models), serializer
-    )
+    return _reserialize(basic_populated_resource_service, serializer)
 
 
 @pytest.fixture
 async def populated_resource_service(
-    resource_service: ResourceServiceInterface,
-    tree3_resource_models,
-    serializer: SerializerInterface,
+    populated_resource_service,
+    serializer: SerializationServiceInterface,
 ):
-    return _reserialize(
-        await _populated_resource_service(resource_service, tree3_resource_models), serializer
-    )
+    return _reserialize(populated_resource_service, serializer)
 
 
 @pytest.fixture
 async def triple_populated_resource_service(
-    resource_service: ResourceServiceInterface,
-    tree1_resource_models,
-    tree2_resource_models,
-    tree3_resource_models,
-    serializer: SerializerInterface,
+    triple_populated_resource_service,
+    serializer: SerializationServiceInterface,
 ):
     return _reserialize(
-        await _triple_populated_resource_service(
-            resource_service, tree1_resource_models, tree2_resource_models, tree3_resource_models
-        ),
+        triple_populated_resource_service,
         serializer,
     )
 
@@ -57,7 +37,7 @@ class TestDeserializedSerializedResourceService(TestResourceService):
 
 
 def _reserialize(
-    resource_service: ResourceServiceInterface, serializer: SerializerInterface
+    resource_service: ResourceServiceInterface, serializer: SerializationServiceInterface
 ) -> ResourceServiceInterface:
-    serialized = serializer.obj_to_pjson(resource_service, ResourceServiceInterface)
-    return serializer.pjson_to_obj(serialized, ResourceServiceInterface)
+    serialized = serializer.to_json(resource_service, ResourceServiceInterface)
+    return serializer.from_json(serialized, ResourceServiceInterface)
